refactor(freeze-updates): extract pure chain helpers into pending-update-chain.ts
mergePendingChains and extractActionsFromChain are pure functions over
React's circular pending-update chain — no shared state, no
side effects. Move them out of freeze-updates.ts into
freeze/pending-update-chain.ts.

freeze-updates.ts: 525 -> 484 (-41). 13 freeze-updates specs green.
